refactor: Move Partition type and friends to their own file
