+++ data_types/src/lib.rs
@@ -24,11 +24,13 @@ mod columns;
 pub use columns::*;
 mod namespace_name;
 pub use namespace_name::*;
+pub mod partition;
+pub use partition::*;
 pub mod partition_template;
 use partition_template::*;
 
 use observability_deps::tracing::warn;
-use schema::{sort::SortKey, TIME_COLUMN_NAME};
+use schema::TIME_COLUMN_NAME;
 use std::{
     borrow::Borrow,
     collections::{BTreeMap, BTreeSet, HashMap},
@@ -37,7 +39,6 @@ use std::{
     mem::{self, size_of_val},
     num::{FpCategory, NonZeroU64},
     ops::{Add, Deref, Sub},
-    sync::Arc,
 };
 use uuid::Uuid;
 
@@ -149,27 +150,6 @@ impl TableId {
 }
 
 impl std::fmt::Display for TableId {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-/// Unique ID for a `Partition`
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, sqlx::Type, sqlx::FromRow)]
-#[sqlx(transparent)]
-pub struct PartitionId(i64);
-
-#[allow(missing_docs)]
-impl PartitionId {
-    pub const fn new(v: i64) -> Self {
-        Self(v)
-    }
-    pub fn get(&self) -> i64 {
-        self.0
-    }
-}
-
-impl std::fmt::Display for PartitionId {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         write!(f, "{}", self.0)
     }
@@ -483,150 +463,6 @@ impl TableSchema {
     /// Return number of columns of the table
     pub fn column_count(&self) -> usize {
         self.columns.column_count()
-    }
-}
-
-/// Defines an partition via an arbitrary string within a table within
-/// a namespace.
-///
-/// Implemented as a reference-counted string, serialisable to
-/// the Postgres VARCHAR data type.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct PartitionKey(Arc<str>);
-
-impl PartitionKey {
-    /// Returns true if this instance of [`PartitionKey`] is backed by the same
-    /// string storage as other.
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        Arc::ptr_eq(&self.0, &other.0)
-    }
-
-    /// Returns underlying string.
-    pub fn inner(&self) -> &str {
-        &self.0
-    }
-}
-
-impl Display for PartitionKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.0)
-    }
-}
-
-impl From<String> for PartitionKey {
-    fn from(s: String) -> Self {
-        assert!(!s.is_empty());
-        Self(s.into())
-    }
-}
-
-impl From<&str> for PartitionKey {
-    fn from(s: &str) -> Self {
-        assert!(!s.is_empty());
-        Self(s.into())
-    }
-}
-
-impl sqlx::Type<sqlx::Postgres> for PartitionKey {
-    fn type_info() -> sqlx::postgres::PgTypeInfo {
-        // Store this type as VARCHAR
-        sqlx::postgres::PgTypeInfo::with_name("VARCHAR")
-    }
-}
-
-impl sqlx::Encode<'_, sqlx::Postgres> for PartitionKey {
-    fn encode_by_ref(
-        &self,
-        buf: &mut <sqlx::Postgres as sqlx::database::HasArguments<'_>>::ArgumentBuffer,
-    ) -> sqlx::encode::IsNull {
-        <&str as sqlx::Encode<sqlx::Postgres>>::encode(&self.0, buf)
-    }
-}
-
-impl sqlx::Decode<'_, sqlx::Postgres> for PartitionKey {
-    fn decode(
-        value: <sqlx::Postgres as sqlx::database::HasValueRef<'_>>::ValueRef,
-    ) -> Result<Self, Box<dyn std::error::Error + 'static + Send + Sync>> {
-        Ok(Self(
-            <String as sqlx::Decode<sqlx::Postgres>>::decode(value)?.into(),
-        ))
-    }
-}
-
-impl sqlx::Type<sqlx::Sqlite> for PartitionKey {
-    fn type_info() -> sqlx::sqlite::SqliteTypeInfo {
-        <String as sqlx::Type<sqlx::Sqlite>>::type_info()
-    }
-}
-
-impl sqlx::Encode<'_, sqlx::Sqlite> for PartitionKey {
-    fn encode_by_ref(
-        &self,
-        buf: &mut <sqlx::Sqlite as sqlx::database::HasArguments<'_>>::ArgumentBuffer,
-    ) -> sqlx::encode::IsNull {
-        <String as sqlx::Encode<sqlx::Sqlite>>::encode(self.0.to_string(), buf)
-    }
-}
-
-impl sqlx::Decode<'_, sqlx::Sqlite> for PartitionKey {
-    fn decode(
-        value: <sqlx::Sqlite as sqlx::database::HasValueRef<'_>>::ValueRef,
-    ) -> Result<Self, Box<dyn std::error::Error + 'static + Send + Sync>> {
-        Ok(Self(
-            <String as sqlx::Decode<sqlx::Sqlite>>::decode(value)?.into(),
-        ))
-    }
-}
-
-/// Data object for a partition. The combination of table and key are unique (i.e. only one record
-/// can exist for each combo)
-#[derive(Debug, Clone, PartialEq, Eq, sqlx::FromRow)]
-pub struct Partition {
-    /// the id of the partition
-    pub id: PartitionId,
-    /// the table the partition is under
-    pub table_id: TableId,
-    /// the string key of the partition
-    pub partition_key: PartitionKey,
-    /// vector of column names that describes how *every* parquet file
-    /// in this [`Partition`] is sorted. The sort_key contains all the
-    /// primary key (PK) columns that have been persisted, and nothing
-    /// else. The PK columns are all `tag` columns and the `time`
-    /// column.
-    ///
-    /// Even though it is possible for both the unpersisted data
-    /// and/or multiple parquet files to contain different subsets of
-    /// columns, the partition's sort_key is guaranteed to be
-    /// "compatible" across all files. Compatible means that the
-    /// parquet file is sorted in the same order as the partition
-    /// sort_key after removing any missing columns.
-    ///
-    /// Partitions are initially created before any data is persisted
-    /// with an empty sort_key. The partition sort_key is updated as
-    /// needed when data is persisted to parquet files: both on the
-    /// first persist when the sort key is empty, as on subsequent
-    /// persist operations when new tags occur in newly inserted data.
-    ///
-    /// Updating inserts new column into the existing order. The order
-    /// of the existing columns relative to each other is NOT changed.
-    ///
-    /// For example, updating `A,B,C` to either `A,D,B,C` or `A,B,C,D`
-    /// is legal. However, updating to `A,C,D,B` is not because the
-    /// relative order of B and C have been reversed.
-    pub sort_key: Vec<String>,
-
-    /// The time at which the newest file of the partition is created
-    pub new_file_at: Option<Timestamp>,
-}
-
-impl Partition {
-    /// The sort key for the partition, if present, structured as a `SortKey`
-    pub fn sort_key(&self) -> Option<SortKey> {
-        if self.sort_key.is_empty() {
-            return None;
-        }
-
-        Some(SortKey::from_columns(self.sort_key.iter().map(|s| &**s)))
     }
 }
